requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0
python-multipart==0.0.6
httpx==0.26.0
pandas==2.1.4
scikit-learn==1.3.2
aiosqlite==0.19.0
pydantic==2.5.3
python-dotenv==1.0.0
